Create bug reproducer
Adapt tests to prove that grouping definitions dont follow
hierarchy of the module but instead they depends on grouping naming.

When we change groupingroot.G1 to groupingroot.G4 we are no more able
to get its definition because its is replaced by previously "rewritten"
G1 grouping which was not defined without this change:

Try to run test without this patch - you can see that we have
definition for G2 but not for G1, instead we have groupingroot.G1!

Signed-off-by: Ivan Hrasko ivan.hrasko@pantheon.tech